chore: move all app package code to new non-internal client package, remove root internal dir, and consolidate types in root types dir
