+++ client/const.go
@@ -1,7 +1,7 @@
 // Copyright (c) Palo Alto Networks, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package app
+package client
 
 const (
 	// Max value for limit in pagination queries.
