+++ cloudonboarding/integration.go
@@ -7,8 +7,8 @@ import (
 	"context"
 	"net/http"
 
-	"github.com/PaloAltoNetworks/cortex-cloud-go/internal/app"
-	"github.com/PaloAltoNetworks/cortex-cloud-go/cloudonboarding/types"
+	"github.com/PaloAltoNetworks/cortex-cloud-go/client"
+	"github.com/PaloAltoNetworks/cortex-cloud-go/types"
 )
 
 // CreateTemplate creates a new Cloud Onboarding Integration Template.
@@ -16,7 +16,7 @@ import (
 // TODO: details
 func (c *Client) CreateIntegrationTemplate(ctx context.Context, input types.CreateIntegrationTemplateRequest) (types.CreateTemplateOrEditIntegrationInstanceResponse, error) {
 	var ans types.CreateTemplateOrEditIntegrationInstanceResponse
-	_, err := c.internalClient.Do(ctx, http.MethodPost, CreateIntegrationTemplateEndpoint, nil, nil, input, &ans, &app.DoOptions{
+	_, err := c.internalClient.Do(ctx, http.MethodPost, CreateIntegrationTemplateEndpoint, nil, nil, input, &ans, &client.DoOptions{
 		RequestWrapperKey:  "request_data",
 		ResponseWrapperKey: "reply",
 	})
@@ -26,7 +26,7 @@ func (c *Client) CreateIntegrationTemplate(ctx context.Context, input types.Crea
 // GetDetails returns the configuration details of the specified integration instance.
 func (c *Client) GetIntegrationInstanceDetails(ctx context.Context, input types.GetIntegrationInstanceRequest) (types.GetIntegrationInstanceResponse, error) {
 	var ans types.GetIntegrationInstanceResponse
-	_, err := c.internalClient.Do(ctx, http.MethodPost, GetIntegrationInstanceDetailsEndpoint, nil, nil, input, &ans, &app.DoOptions{
+	_, err := c.internalClient.Do(ctx, http.MethodPost, GetIntegrationInstanceDetailsEndpoint, nil, nil, input, &ans, &client.DoOptions{
 		RequestWrapperKey:  "request_data",
 		ResponseWrapperKey: "reply",
 	})
@@ -35,7 +35,7 @@ func (c *Client) GetIntegrationInstanceDetails(ctx context.Context, input types.
 
 func (c *Client) ListIntegrationInstances(ctx context.Context, input types.ListIntegrationInstancesRequest) (types.ListIntegrationInstancesResponseWrapper, error) {
 	var ans types.ListIntegrationInstancesResponseWrapper
-	_, err := c.internalClient.Do(ctx, http.MethodPost, ListIntegrationInstancesEndpoint, nil, nil, input, &ans, &app.DoOptions{
+	_, err := c.internalClient.Do(ctx, http.MethodPost, ListIntegrationInstancesEndpoint, nil, nil, input, &ans, &client.DoOptions{
 		RequestWrapperKey:  "request_data",
 		ResponseWrapperKey: "reply",
 	})
@@ -43,7 +43,7 @@ func (c *Client) ListIntegrationInstances(ctx context.Context, input types.ListI
 }
 func (c *Client) EditIntegrationInstance(ctx context.Context, input types.EditIntegrationInstanceRequest) (types.CreateTemplateOrEditIntegrationInstanceResponse, error) {
 	var ans types.CreateTemplateOrEditIntegrationInstanceResponse
-	_, err := c.internalClient.Do(ctx, http.MethodPost, EditIntegrationInstanceEndpoint, nil, nil, input, &ans, &app.DoOptions{
+	_, err := c.internalClient.Do(ctx, http.MethodPost, EditIntegrationInstanceEndpoint, nil, nil, input, &ans, &client.DoOptions{
 		RequestWrapperKey:  "request_data",
 		ResponseWrapperKey: "reply",
 	})
@@ -52,21 +52,21 @@ func (c *Client) EditIntegrationInstance(ctx context.Context, input types.EditIn
 }
 
 func (c *Client) EnableIntegrationInstances(ctx context.Context, instanceIDs []string) error {
-	_, err := c.internalClient.Do(ctx, http.MethodPost, EnableOrDisableIntegrationInstancesEndpoint, nil, nil, types.EnableOrDisableIntegrationInstancesRequest{ IDs: instanceIDs, Enable: true }, nil, &app.DoOptions{
+	_, err := c.internalClient.Do(ctx, http.MethodPost, EnableOrDisableIntegrationInstancesEndpoint, nil, nil, types.EnableOrDisableIntegrationInstancesRequest{ IDs: instanceIDs, Enable: true }, nil, &client.DoOptions{
 		RequestWrapperKey: "request_data",
 	})
 	return err
 }
 
 func (c *Client) DisableIntegrationInstances(ctx context.Context, instanceIDs []string) error {
-	_, err := c.internalClient.Do(ctx, http.MethodPost, EnableOrDisableIntegrationInstancesEndpoint, nil, nil, types.EnableOrDisableIntegrationInstancesRequest{ IDs: instanceIDs, Enable: false }, nil, &app.DoOptions{
+	_, err := c.internalClient.Do(ctx, http.MethodPost, EnableOrDisableIntegrationInstancesEndpoint, nil, nil, types.EnableOrDisableIntegrationInstancesRequest{ IDs: instanceIDs, Enable: false }, nil, &client.DoOptions{
 		RequestWrapperKey: "request_data",
 	})
 	return err
 }
 
 func (c *Client) DeleteIntegrationInstances(ctx context.Context, instanceIDs []string) error {
-	_, err := c.internalClient.Do(ctx, http.MethodPost, DeleteIntegrationInstancesEndpoint, nil, nil, types.DeleteIntegrationInstanceRequest{ IDs: instanceIDs }, nil, &app.DoOptions{
+	_, err := c.internalClient.Do(ctx, http.MethodPost, DeleteIntegrationInstancesEndpoint, nil, nil, types.DeleteIntegrationInstanceRequest{ IDs: instanceIDs }, nil, &client.DoOptions{
 		RequestWrapperKey: "request_data",
 	})
 	return err
