+++ go.work
@@ -3,12 +3,11 @@ go 1.25.0
 use (
 	./api
 	./appsec
+	./client
 	./cloudonboarding
 	./cwp
 	./enums
 	./errors
-	./internal/app
-	./internal/util
 	./log
 	./platform
 	./types
