--- internal/app/go.mod
@@ -1,23 +1,0 @@
-module github.com/PaloAltoNetworks/cortex-cloud-go/internal/app
-
-go 1.25.0
-
-require (
-	github.com/PaloAltoNetworks/cortex-cloud-go/api v0.0.6
-	github.com/PaloAltoNetworks/cortex-cloud-go/errors v0.0.6
-	github.com/PaloAltoNetworks/cortex-cloud-go/log v0.0.6
-	github.com/stretchr/testify v1.10.0
-)
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/fatih/color v1.13.0 // indirect
-	github.com/google/go-cmp v0.6.0 // indirect
-	github.com/hashicorp/go-hclog v1.5.0 // indirect
-	github.com/hashicorp/terraform-plugin-log v0.9.0 // indirect
-	github.com/mattn/go-colorable v0.1.12 // indirect
-	github.com/mattn/go-isatty v0.0.14 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	golang.org/x/sys v0.30.0 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
