--- internal/util/go.mod
@@ -1,3 +1,0 @@
-module github.com/PaloAltoNetworks/cortex-cloud-go/internal/util
-
-go 1.25.0
