+++ platform/asset_group.go
@@ -5,8 +5,8 @@ import (
 	"net/http"
 	"strconv"
 
-	"github.com/PaloAltoNetworks/cortex-cloud-go/internal/app"
-	"github.com/PaloAltoNetworks/cortex-cloud-go/platform/types"
+	"github.com/PaloAltoNetworks/cortex-cloud-go/client"
+	"github.com/PaloAltoNetworks/cortex-cloud-go/types"
 )
 
 // CreateAssetGroup creates a new asset group.
@@ -14,7 +14,7 @@ import (
 // TODO: make sure that the SEARCH_FIELD value is forced to uppercase
 func (c *Client) CreateAssetGroup(ctx context.Context, req types.CreateOrUpdateAssetGroupRequest) (types.CreateAssetGroupResponseData, error) {
 	var respWrapper types.CreateAssetGroupResponseWrapper
-	_, err := c.internalClient.Do(ctx, http.MethodPost, CreateAssetGroupEndpoint, nil, nil, types.CreateOrUpdateAssetGroupRequestWrapper{AssetGroup: req}, &respWrapper, &app.DoOptions{
+	_, err := c.internalClient.Do(ctx, http.MethodPost, CreateAssetGroupEndpoint, nil, nil, types.CreateOrUpdateAssetGroupRequestWrapper{AssetGroup: req}, &respWrapper, &client.DoOptions{
 		RequestWrapperKey:  "request_data",
 		ResponseWrapperKey: "reply",
 	})
@@ -24,7 +24,7 @@ func (c *Client) CreateAssetGroup(ctx context.Context, req types.CreateOrUpdateA
 // ListAssetGroups retrieves a list of asset groups.
 func (c *Client) ListAssetGroups(ctx context.Context, req types.ListAssetGroupsRequest) (types.ListAssetGroupsResponse, error) {
 	var resp types.ListAssetGroupsResponse
-	_, err := c.internalClient.Do(ctx, http.MethodPost, ListAssetGroupsEndpoint, nil, nil, req, &resp, &app.DoOptions{
+	_, err := c.internalClient.Do(ctx, http.MethodPost, ListAssetGroupsEndpoint, nil, nil, req, &resp, &client.DoOptions{
 		RequestWrapperKey:  "request_data",
 		ResponseWrapperKey: "reply",
 	})
@@ -35,7 +35,7 @@ func (c *Client) ListAssetGroups(ctx context.Context, req types.ListAssetGroupsR
 func (c *Client) UpdateAssetGroup(ctx context.Context, groupID int, req types.CreateOrUpdateAssetGroupRequest) (types.GenericAssetGroupsResponseData, error) {
 	var respWrapper types.GenericAssetGroupsResponseWrapper
 	pathParams := &[]string{strconv.Itoa(groupID)}
-	_, err := c.internalClient.Do(ctx, http.MethodPost, UpdateAssetGroupEndpoint, pathParams, nil, types.CreateOrUpdateAssetGroupRequestWrapper{AssetGroup: req}, &respWrapper, &app.DoOptions{
+	_, err := c.internalClient.Do(ctx, http.MethodPost, UpdateAssetGroupEndpoint, pathParams, nil, types.CreateOrUpdateAssetGroupRequestWrapper{AssetGroup: req}, &respWrapper, &client.DoOptions{
 		RequestWrapperKey:  "request_data",
 		ResponseWrapperKey: "reply",
 	})
@@ -46,7 +46,7 @@ func (c *Client) UpdateAssetGroup(ctx context.Context, groupID int, req types.Cr
 func (c *Client) DeleteAssetGroup(ctx context.Context, groupID int) (types.GenericAssetGroupsResponseData, error) {
 	var respWrapper types.GenericAssetGroupsResponseWrapper
 	pathParams := &[]string{strconv.Itoa(groupID)}
-	_, err := c.internalClient.Do(ctx, http.MethodPost, DeleteAssetGroupEndpoint, pathParams, nil, nil, &respWrapper, &app.DoOptions{
+	_, err := c.internalClient.Do(ctx, http.MethodPost, DeleteAssetGroupEndpoint, pathParams, nil, nil, &respWrapper, &client.DoOptions{
 		ResponseWrapperKey: "reply",
 	})
 	return respWrapper.Data, err
