+++ types/go.mod
@@ -1,3 +1,21 @@
 module github.com/PaloAltoNetworks/cortex-cloud-go/types
 
 go 1.25.0
+
+replace github.com/PaloAltoNetworks/cortex-cloud-go/api => ../api
+
+replace github.com/PaloAltoNetworks/cortex-cloud-go/appsec => ../appsec
+
+replace github.com/PaloAltoNetworks/cortex-cloud-go/client => ../client
+
+replace github.com/PaloAltoNetworks/cortex-cloud-go/cloudonboarding => ../cloudonboarding
+
+replace github.com/PaloAltoNetworks/cortex-cloud-go/cwp => ../cwp
+
+replace github.com/PaloAltoNetworks/cortex-cloud-go/enums => ../enums
+
+replace github.com/PaloAltoNetworks/cortex-cloud-go/errors => ../errors
+
+replace github.com/PaloAltoNetworks/cortex-cloud-go/log => ../log
+
+replace github.com/PaloAltoNetworks/cortex-cloud-go/platform => ../platform
