+++ types/option.go
@@ -1,0 +1,6 @@
+package types
+//
+//type DoOptions struct {
+//	RequestWrapperKey  string
+//	ResponseWrapperKey string
+//}
